enhance replication error management
Split error management in two categories: retryable and permanent
failures. The choice is currently based on the 'retryable' field of
the error sent back by the AWS SDK, but it could be refined later.

 - retryable errors get retried indefinitely with an exponential
   back-off algorithm (using 'backo' open-source module, it's simple
   but non-trivial to do properly so makes sense to use a third-party
   module IMO). We may later think about giving up after some (long
   enough) time without success and set the replication status to
   FAILED, not sure we want to though, if the errors are really
   retryable they should eventually succeed.

 - permanent errors lead to the status FAILED being written on the
   source object version.

 - failures to write the replication status also lead to retries.

Tests should be written, but they will be in a later PR as the code
will need instrumentation that will likely conflict with other pending
PRs.